BTFS-1590:refactor: move storage sub commands into different files
