+++ core/commands/store/hosts/hosts.go
@@ -1,0 +1,118 @@
+package hosts
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/TRON-US/go-btfs/core"
+	"github.com/TRON-US/go-btfs/core/commands/cmdenv"
+	"github.com/TRON-US/go-btfs/core/commands/storage"
+	"github.com/TRON-US/go-btfs/core/hub"
+
+	cmds "github.com/TRON-US/go-btfs-cmds"
+	hubpb "github.com/tron-us/go-btfs-common/protos/hub"
+)
+
+const (
+	hostInfoModeOptionName = "host-info-mode"
+	hostSyncModeOptionName = "host-sync-mode"
+)
+
+var StorageHostsCmd = &cmds.Command{
+	Helptext: cmds.HelpText{
+		Tagline:          "Interact with information on hosts.",
+		ShortDescription: `Allows interaction with information on hosts. Host information is synchronized from btfs-hub and saved in local datastore.`,
+	},
+	Subcommands: map[string]*cmds.Command{
+		"info": storageHostsInfoCmd,
+		"sync": storageHostsSyncCmd,
+	},
+}
+
+var storageHostsInfoCmd = &cmds.Command{
+	Helptext: cmds.HelpText{
+		Tagline: "Display saved host information.",
+		ShortDescription: `
+This command displays saved information from btfs-hub under multiple modes.
+Each mode ranks hosts based on its criteria and is randomized based on current node location.
+
+Mode options include:` + hub.AllModeHelpText,
+	},
+	Options: []cmds.Option{
+		cmds.StringOption(hostInfoModeOptionName, "m", "Hosts info showing mode. Default: mode set in config option Experimental.HostsSyncMode."),
+	},
+	Run: func(req *cmds.Request, res cmds.ResponseEmitter, env cmds.Environment) error {
+		cfg, err := cmdenv.GetConfig(env)
+		if err != nil {
+			return err
+		}
+		if !cfg.Experimental.StorageClientEnabled {
+			return fmt.Errorf("storage client api not enabled")
+		}
+
+		mode, ok := req.Options[hostInfoModeOptionName].(string)
+		if !ok {
+			mode = cfg.Experimental.HostsSyncMode
+		}
+
+		n, err := cmdenv.GetNode(env)
+		if err != nil {
+			return err
+		}
+
+		nodes, err := storage.GetHostsFromDatastore(req.Context, n, mode, 0)
+		if err != nil {
+			return err
+		}
+
+		return cmds.EmitOnce(res, &HostInfoRes{nodes})
+	},
+	Type: HostInfoRes{},
+}
+
+type HostInfoRes struct {
+	Nodes []*hubpb.Host
+}
+
+var storageHostsSyncCmd = &cmds.Command{
+	Helptext: cmds.HelpText{
+		Tagline: "Synchronize host information from btfs-hub.",
+		ShortDescription: `
+This command synchronizes information from btfs-hub using multiple modes.
+Each mode ranks hosts based on its criteria and is randomized based on current node location.
+
+Mode options include:` + hub.AllModeHelpText,
+	},
+	Options: []cmds.Option{
+		cmds.StringOption(hostSyncModeOptionName, "m", "Hosts syncing mode. Default: mode set in config option Experimental.HostsSyncMode."),
+	},
+	Run: func(req *cmds.Request, res cmds.ResponseEmitter, env cmds.Environment) error {
+		cfg, err := cmdenv.GetConfig(env)
+		if err != nil {
+			return err
+		}
+		if !cfg.Experimental.StorageClientEnabled {
+			return fmt.Errorf("storage client api not enabled")
+		}
+
+		mode, ok := req.Options[hostSyncModeOptionName].(string)
+		if !ok {
+			mode = cfg.Experimental.HostsSyncMode
+		}
+
+		n, err := cmdenv.GetNode(env)
+		if err != nil {
+			return err
+		}
+
+		return SyncHosts(req.Context, n, mode)
+	},
+}
+
+func SyncHosts(ctx context.Context, node *core.IpfsNode, mode string) error {
+	nodes, err := hub.QueryHosts(ctx, node, mode)
+	if err != nil {
+		return err
+	}
+	return storage.SaveHostsIntoDatastore(ctx, node, mode, nodes)
+}
