+++ core/commands/store/upload/init.go
@@ -1,0 +1,1 @@
+package upload
