+++ core/commands/store/upload/recvcontract.go
@@ -1,0 +1,1 @@
+package upload
