+++ core/commands/store/upload/status.go
@@ -1,0 +1,1 @@
+package upload
